+++ R/print.trollsim.R
@@ -17,40 +17,25 @@ NULL
 #' @export
 #' @rdname print.trollsim
 setMethod("print", "trollsim", function(x, ...) {
-  pars <- x@inputs$global$value
-  names(pars) <- x@inputs$global$param
-  cat("Object of class :", class(x)[1], "\n\n")
-  cat("Name :", x@name, "\n\n")
-  cat(
-    "2D discrete network: horizontal step = ",
-    pars["NV"], "m, one tree per 1 m^2 \n"
+  cat("Object of class :", class(x)[1], "\n")
+  cat("Name :", x@name, "\n")
+  cat("Path :", x@path, "\n")
+  cat("Forest :", as.logical(x@parameters["forest"]), "\n")
+  cat("Random :", as.logical(x@parameters["random"]), "\n\n")
+  cat("2D discrete network: horizontal step = ",
+      x@parameters["NV"], "m, one tree per 1 m^2 \n"
   )
   cat(
-    "Number of sites      : ", pars["cols"],
-    "x", pars["rows"], "\n"
+    "Number of sites      : ", x@parameters["cols"],
+    "x", x@parameters["rows"], "\n"
   )
-  cat("Number of iterations : ", pars["nbiter"], "\n")
-  cat("Duration of timestep : ", 1 / pars["iterperyear"] * 365, "days \n")
+  cat("Number of iterations : ", x@parameters["nbiter"], "\n")
+  cat("Duration of timestep : ", 1 / x@parameters["iterperyear"] * 365, "days \n")
   cat("Number of Species    : ", nrow(x@inputs$species), "\n\n")
 })
 
 #' @export
 #' @rdname print.trollsim
 setMethod("show", "trollsim", function(object) {
-  x <- object
-  pars <- x@inputs$global$value
-  names(pars) <- x@inputs$global$param
-  cat("Object of class :", class(x)[1], "\n\n")
-  cat("Name :", x@name, "\n\n")
-  cat(
-    "2D discrete network: horizontal step = ",
-    pars["NV"], "m, one tree per 1 m^2 \n"
-  )
-  cat(
-    "Number of sites      : ", pars["cols"],
-    "x", pars["rows"], "\n"
-  )
-  cat("Number of iterations : ", pars["nbiter"], "\n")
-  cat("Duration of timestep : ", 1 / pars["iterperyear"] * 365, "days \n")
-  cat("Number of Species    : ", nrow(x@inputs$species), "\n\n")
+  print(object)
 })